Use `u32` for aggregate version
sqlx can only handle 32-bit integers in combination with sqlite. Even a
u32 gives us a lot of room (more than we'd hopefully ever need) for how
many events we can apply.

To avoid casting when interacting with the DB (to come in a future patch)
we change the version to use a u32.

+++ daemon/src/projection.rs
@@ -215,7 +215,7 @@ struct Aggregated {
     rollover_state: Option<ProtocolNegotiationState>,
     settlement_state: Option<ProtocolNegotiationState>,
 
-    version: u64,
+    version: u32,
     creation_timestamp: Timestamp,
 }
